+++ IISU/ImplementedStoreTypes/WinIIS/Management.cs
@@ -25,15 +25,13 @@ using Microsoft.Extensions.Logging;
 using Microsoft.PowerShell.Commands;
 using Newtonsoft.Json;
 
-namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore.WinIIS
+namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore.IISU
 {
     public class Management : WinCertJobTypeBase, IManagementJobExtension
     {
         private ILogger _logger;
 
         public string ExtensionName => string.Empty;
-
-        private string _thumbprint = string.Empty;
 
         private Runspace myRunspace;
 
@@ -49,7 +47,7 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore.WinIIS
                 _logger = LogHandler.GetClassLogger<Management>();
                 _logger.MethodEntry();
 
-                _logger.LogTrace($"Job Configuration: {JsonConvert.SerializeObject(config)}");
+                _logger.LogTrace(JobConfigurationParser.ParseManagementJobConfiguration(config));
 
                 string serverUserName = PAMUtilities.ResolvePAMField(_resolver, _logger, "Server UserName", config.ServerUsername);
                 string serverPassword = PAMUtilities.ResolvePAMField(_resolver, _logger, "Server Password", config.ServerPassword);
@@ -139,8 +137,6 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore.WinIIS
         {
             _logger.LogTrace("Before Remove Certificate...");
 
-            string certificateContents = config.JobCertificate.Contents;
-            string privateKeyPassword = config.JobCertificate.PrivateKeyPassword;
             string storePath = config.CertificateStoreDetails.StorePath;
             long jobNumber = config.JobHistoryId;