+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-2.0.1
+2.1.0
 * Fixed issue that was occuring during renewal when there were bindings outside of http and https like net.tcp
 * Added PAM registration/initialization documentation in README.md
+* Resolved Null HostName error 
+* Added WinCert Cert Store Type
+* Added custom property parser to not show any passwords
+* Removed any password references in trace logs and output settings in JSON format
 
 2.0.0
 * Add support for reenrollment jobs (On Device Key Generation) with the ability to specify a cryptographic provider. Specification of cryptographic provider allows HSM (Hardware Security Module) use.